chore(deps): fix mybatis-plus broken API after update from 3.4.3.1 to 3.4.3.2

+++ ji-admin/src/main/java/com/power4j/ji/admin/modules/sys/service/SysDictItemService.java
@@ -45,6 +45,6 @@ public interface SysDictItemService extends CrudService<SysDictItemDTO, SysDictI
 	 * @param dictId
 	 * @return
 	 */
-	int countDictItemValue(String value, @Nullable Long ignoreId, Serializable dictId);
+	long countDictItemValue(String value, @Nullable Long ignoreId, Serializable dictId);
 
 }
