+++ ji-admin/src/main/java/com/power4j/ji/admin/modules/sys/service/SysDictService.java
@@ -36,7 +36,7 @@ public interface SysDictService extends CrudService<SysDictDTO, SysDict> {
 	 * @param ignoreId
 	 * @return
 	 */
-	int countDictCode(String code, @Nullable Long ignoreId);
+	long countDictCode(String code, @Nullable Long ignoreId);
 
 	/**
 	 * 查询
