+++ ji-admin/src/main/java/com/power4j/ji/admin/modules/sys/service/SysRoleService.java
@@ -37,7 +37,7 @@ public interface SysRoleService extends CrudService<SysRoleDTO, SysRole> {
 	 * @param ignoreId 排除的ID
 	 * @return
 	 */
-	int countRoleCode(String code, @Nullable Long ignoreId);
+	long countRoleCode(String code, @Nullable Long ignoreId);
 
 	/**
 	 * 角色列表
