+++ ji-admin/src/main/java/com/power4j/ji/admin/modules/sys/service/impl/SysDictItemServiceImpl.java
@@ -57,7 +57,7 @@ public class SysDictItemServiceImpl extends AbstractCrudService<SysDictItemMappe
 	}
 
 	@Override
-	public int countDictItemValue(String value, Long ignoreId, Serializable dictId) {
+	public long countDictItemValue(String value, Long ignoreId, Serializable dictId) {
 		Wrapper<SysDictItem> wrapper = Wrappers.<SysDictItem>lambdaQuery().eq(SysDictItem::getDictId, dictId)
 				.eq(SysDictItem::getValue, value).ne(null != ignoreId, SysDictItem::getId, ignoreId);
 		return getBaseMapper().selectCount(wrapper);
