+++ ji-admin/src/main/java/com/power4j/ji/admin/modules/sys/service/impl/SysOrgServiceImpl.java
@@ -55,7 +55,7 @@ public class SysOrgServiceImpl
 	}
 
 	@Override
-	public int countOrgCode(String code, @Nullable Long ignoreId) {
+	public long countOrgCode(String code, @Nullable Long ignoreId) {
 		return countByLambdaColumn(SysOrg::getCode, code, ignoreId);
 	}
