+++ ji-admin/src/main/java/com/power4j/ji/admin/modules/sys/service/impl/SysParamServiceImpl.java
@@ -45,7 +45,7 @@ public class SysParamServiceImpl extends AbstractCrudService<SysParamMapper, Sys
 		implements SysParamService {
 
 	@Override
-	public int countParamKey(String key, @Nullable Long ignoreId) {
+	public long countParamKey(String key, @Nullable Long ignoreId) {
 		return countByLambdaColumn(SysParam::getParamKey, key, ignoreId);
 	}
