+++ ji-admin/src/main/java/com/power4j/ji/admin/modules/ureport/service/UrFileService.java
@@ -39,7 +39,7 @@ public interface UrFileService extends CrudService<UrFileDTO, UrData> {
 	 * @param ignoreId 排除的ID
 	 * @return
 	 */
-	int countFileName(String file, @Nullable Long ignoreId);
+	long countFileName(String file, @Nullable Long ignoreId);
 
 	/**
 	 * 查询
